move getRemindersForMonth() to AccountHelper
