+++ app/Helpers/AccountHelper.php
@@ -74,6 +74,34 @@ class AccountHelper
     }
 
     /**
+     * Get the reminders for the month given in parameter.
+     * - 0 means current month
+     * - 1 means month+1
+     * - 2 means month+2...
+     *
+     * @param Account $account
+     * @param int $month
+     */
+    public static function getUpcomingRemindersForMonth(Account $account, int $month)
+    {
+        $startOfMonth = now(DateHelper::getTimezone())->addMonthsNoOverflow($month)->startOfMonth();
+
+        // don't get reminders for past events:
+        if ($startOfMonth->isPast()) {
+            $startOfMonth = now(DateHelper::getTimezone());
+        }
+
+        $endOfMonth = now(DateHelper::getTimezone())->addMonthsNoOverflow($month)->endOfMonth();
+
+        return $account->reminderOutboxes()
+            ->with(['reminder', 'reminder.contact'])
+            ->whereBetween('planned_date', [$startOfMonth, $endOfMonth])
+            ->where('nature', 'reminder')
+            ->orderBy('planned_date', 'asc')
+            ->get();
+    }
+
+    /**
      * Get the number of activities grouped by year.
      *
      * @param Account $account
