+++ tests/Unit/Models/AccountTest.php
@@ -351,48 +351,6 @@ class AccountTest extends FeatureTestCase
     }
 
     /** @test */
-    public function get_reminders_for_month_returns_no_reminders()
-    {
-        $user = $this->signIn();
-
-        $account = $user->account;
-
-        Carbon::setTestNow(Carbon::create(2017, 1, 1));
-        factory(Reminder::class, 3)->create(['account_id' => $account->id]);
-
-        // check if there are reminders for the month of March
-        $this->assertEquals(
-            0,
-            $account->getRemindersForMonth(3)->count()
-        );
-    }
-
-    /** @test */
-    public function get_reminders_for_month_returns_reminders_for_given_month()
-    {
-        $user = $this->signIn();
-
-        $account = $user->account;
-
-        Carbon::setTestNow(Carbon::create(2017, 1, 1));
-
-        // add 3 reminders for the month of March
-        for ($i = 0; $i < 3; $i++) {
-            $reminder = factory(Reminder::class)->create([
-                'account_id' => $account->id,
-                'initial_date' => '2017-03-03 00:00:00',
-            ]);
-
-            $reminder->schedule($user);
-        }
-
-        $this->assertEquals(
-            3,
-            $account->getRemindersForMonth(2)->count()
-        );
-    }
-
-    /** @test */
     public function it_gets_the_id_of_the_subscribed_plan()
     {
         config([
